app/requirements.txt
streamlit==1.37.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
plotly==5.23.0
joblib==1.4.2
scikit-learn==1.5.1
